cleanup code & add code doc

+++ sources/Application/Views/SampleEditorView.h
@@ -83,7 +83,7 @@ private:
   collectSampleUsers(int sampleIndex) const;
   void retargetSampleUsers(
       const etl::vector<SampleInstrument *, MAX_INSTRUMENT_COUNT> &users,
-      int newIndex);
+      uint16_t newIndex);
   void loadSample(const etl::string<MAX_INSTRUMENT_FILENAME_LENGTH> path,
                   bool isProjectSampleFile);
   bool reloadEditedSample();
